Setup initial github action (#1)
Setup an initial github action for building and pushing the docker image. The test_notebooks are used for test, with the exception of the cadquery one that seems to need more changes. Another one importing all the packages was added. No output validation is done, but an exception will lead to a test failure.

Remove the tag from the repo, replace it with docker/metadata-actio from the github context.

To make the build and test work, multiple changes are done:

 - Explicitly add pip in the dependencies, conda was complaining about it in loop during docker build.
 - The pyvista wheel that was included was triggering import failures (missing pyvista.jupyter). Replace it with the latest version
 - Minor update to the test example to allow running with an X server
 - Replace ipyvtk_simple with ipyvtklink, that was throwing a deprecated error